Authorization With Policies

+++ app/Http/Controllers/ThreadController.php
@@ -106,6 +106,8 @@ class ThreadController extends Controller
      */
     public function destroy(Thread $thread)
     {
+        $this->authorize('update', $thread);
+
         $thread->delete();
 
         return redirect('threads');
